+++ include/pbcopper/parallel/FireAndForget.h
@@ -1,5 +1,3 @@
-// Author: Lance Hepler & Armin Töpfer
-
 #ifndef PBCOPPER_PARALLEL_FIREANDFORGET_H
 #define PBCOPPER_PARALLEL_FIREANDFORGET_H
 
@@ -8,10 +6,13 @@
 #include <atomic>
 #include <condition_variable>
 #include <cstddef>
+#include <cstdint>
 #include <exception>
+#include <functional>
 #include <future>
 #include <mutex>
 #include <queue>
+#include <stdexcept>
 
 #include <boost/optional.hpp>
 
@@ -25,7 +26,12 @@ private:
 
 public:
     FireAndForget(const size_t size, const size_t mul = 2)
-        : exc{nullptr}, sz{size * mul}, abort{false}, thrown{false}
+        : exc{nullptr}
+        , numThreads{size}
+        , sz{size * mul}
+        , abort{false}
+        , thrown{false}
+        , acceptingJobs{true}
     {
         for (size_t i = 0; i < size; ++i) {
             threads.emplace_back(std::thread([this]() {
@@ -62,6 +68,11 @@ public:
     template <typename F, typename... Args>
     void ProduceWith(F&& f, Args&&... args)
     {
+        if (!acceptingJobs) {
+            throw std::runtime_error(
+                "FireAndForget error: Cannot dispatch jobs to finalized thread pool!");
+        }
+
         std::packaged_task<void()> task{std::bind(std::forward<F>(f), std::forward<Args>(args)...)};
 
         {
@@ -87,6 +98,7 @@ public:
 
     void Finalize()
     {
+        acceptingJobs = false;
         {
             std::lock_guard<std::mutex> g(m);
             // Push boost::none to signal that there are no further tasks
@@ -106,6 +118,8 @@ public:
             std::rethrow_exception(exc);
         }
     }
+
+    size_t NumThreads() const { return numThreads; }
 
 private:
     TTask PopTask()
@@ -135,10 +149,23 @@ private:
     std::condition_variable pushed;
     std::exception_ptr exc;
     std::mutex m;
+    size_t numThreads;
     size_t sz;
     std::atomic_bool abort;
     std::atomic_bool thrown;
+    std::atomic_bool acceptingJobs;
 };
+
+///
+/// \brief Use an existing FireAndForget to dispatch [0, numEntries) callbacks.
+///        Returns after all dispatched jobs finished.
+///
+/// \param faf         reference to FaF, nullptr allowed
+/// \param numEntries  number of submissions
+/// \param callback    function to be dispatched to FaF
+///
+void Dispatch(Parallel::FireAndForget* faf, int32_t numEntries,
+              const std::function<void(int32_t)>& callback);
 
 }  // namespace Parallel
 }  // namespace PacBio